requirements.txt
asgiref==3.6.0
dj-database-url==0.5.0
Django==3.2
django-countries==7.5.1
django-crispy-forms==2.0
django-shortcuts==1.6
Pillow==9.5.0
psycopg2==2.9.6
pytz==2023.3
sqlparse==0.4.3
